feat: DTO 생성 및 가짜 데이터 생성

+++ capstone_backend/src/main/java/com/example/capstone_backend/domain/user/entity/UserInfo.java
@@ -36,16 +36,16 @@ public class UserInfo {
     private String profile;
 
     @Column
-    private Float height;
+    private Double height;
 
     @Column
-    private Float weight;
+    private Double weight;
 
     @Column
-    private Float muscleMass;
+    private Double muscleMass;
 
     @Column
-    private Float fatMass;
+    private Double fatMass;
 
     @CreatedDate
     @Column
